server/requirements.txt
fastapi==0.115.5
uvicorn==0.32.1
PyYAML==6.0.2
httpx==0.27.2
unsloth-zoo==2024.10.2
